+++ src/components/game/Game.module.scss
@@ -1,4 +1,4 @@
-@import "../_theme.scss";
+@import "../../_theme.scss";
 
 .container {
   display: flex;
