+++ src/index.tsx
@@ -4,7 +4,7 @@ import 'rc-tooltip/assets/bootstrap_white.css'
 import React from 'react'
 import ReactDOM from 'react-dom'
 import 'react-splitter-layout/lib/index.css'
-import Game from './components/Game'
+import Game from './components/game/Game'
 import Worksheet from './components/worksheet/Worksheet'
 import './index.scss'
 import * as serviceWorker from './serviceWorker'
